% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aController}
\alias{aController}
\title{Constructor for a Controller proto objects}
\usage{
aController(...)
}
\arguments{
\item{...}{passed to proto method for}
}
\value{
returns the Controller object
}
\description{
Simply provides a more typical calling interface for the Controller proto object
}

