% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Adapter}
\alias{Adapter}
\title{Trait for Adapter object
An adapter is a simple controller connecting one model property with a widget in a view
by default the adapter synchronizes changes}
\format{\preformatted{proto object 
 $ handler_user_data: NULL 
 $ class            : chr [1:3] "Adapter" "Controller" "TraitR" 
 $ update_from_model:function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 179 49 190 29 49 29 179 190 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266cf1948>  
 $ init             :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 228 34 239 29 34 29 228 239 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266cf1948>  
 $ remove_view      :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 224 41 227 29 41 29 224 227 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266cf1948>  
 $ property         : NULL 
 $ view_widget_name : NULL 
 $ update_from_view :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 192 48 222 29 48 29 192 222 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266cf1948>  
 $ add_handler_name : chr "addHandlerChanged" 
 parent: proto object 
 .. parent: proto object 
}}
\usage{
Adapter
}
\description{
Trait for Adapter object
An adapter is a simple controller connecting one model property with a widget in a view
by default the adapter synchronizes changes
}
\keyword{datasets}

