% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDB.R
\name{addTree}
\alias{addTree}
\title{Function to add a new phylogenetic tree to the local database.}
\usage{
addTree(treeFile, treeFileFormat)
}
\arguments{
\item{treeFile}{String. Relative path of the file containing the tree you wish to add.}

\item{treeFileFormat}{String. The format of the dataset file. Accepts: newick, nexus, nexml, phyloxml, cdao.}
}
\description{
Function to add a new phylogenetic tree to the local database.
}
