% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aModel}
\alias{aModel}
\title{Constructor for a Model proto objects}
\usage{
aModel(...)
}
\arguments{
\item{...}{passed to proto method for}
}
\value{
returns the Model object. Call \code{obj$show_help()} to view its methods and properties.
}
\description{
Simply provides a more typical calling interface for the Model proto object
}

