% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aView}
\alias{aView}
\title{Constructor for a View proto object}
\usage{
aView(...)
}
\arguments{
\item{...}{passed to proto method for}
}
\value{
Returns the View object. Call \code{obj$show_help()} to view its methods and properties.
}
\description{
Simply provides a more typical calling interface for the View proto object
}

