% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{BaseTrait}
\alias{BaseTrait}
\title{Base Trait to place common properties and methods}
\format{\preformatted{proto object 
 $ traitr         : logi TRUE 
 $ do_call        :function (., fun, lst = list())   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 231 31 236 21 31 21 231 236 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ show_help      :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 269 33 297 21 33 21 269 297 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ list_objects   :function (., all.names = FALSE)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 153 35 177 20 35 20 153 177 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ append         :function (., name, value, key)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 68 28 79 21 28 21 68 79 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ new            :function (., ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 44 27 48 21 27 21 44 48 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ list_properties:function (., all.names = FALSE)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 186 36 186 100 36 100 186 186 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ class          : chr "TraitR" 
 $ has_slot       :function (., key)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 102 29 105 20 29 20 102 105 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ next_method    :function (., meth_name)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 199 34 222 20 34 20 199 222 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ has_local_slot :function (., key)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 109 35 111 20 35 20 109 111 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ create_doc_list:function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 239 39 263 21 39 21 239 263 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ add_class      :function (., newclass)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 37 33 37 85 33 85 37 37 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ get_slot       :function (., key)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 118 29 123 20 29 20 118 123 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ set_slot       :function (., key, value, initialize = TRUE)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 140 29 147 20 29 20 140 147 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ get_local_slot :function (., key)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 128 37 133 20 37 20 128 133 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ list_methods   :function (., all.names = FALSE)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 191 33 191 94 33 94 191 191 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ is             :function (., class = NULL)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 89 26 94 21 26 21 89 94 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
 $ assign_if_null :function (., key, value)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 55 35 59 20 35 20 55 59 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266aff340>  
}}
\usage{
BaseTrait
}
\description{
Base Trait to place common properties and methods
}
\keyword{datasets}

