% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{BooleanEditor}
\alias{BooleanEditor}
\title{Trait for Editor for TRUE/FALSE selection}
\format{\preformatted{proto object 
 $ class          : chr [1:4] "BooleanEditor" "Editor" "View" "TraitR" 
 $ editor_name    : chr "gcombobox" 
 $ make_ui_compact:function (., container, attr = .$attr, context, ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 380 47 383 31 47 31 380 383 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2651fc120>  
 $ make_ui        :function (., container, attr = .$attr, context, ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 376 39 379 31 39 31 376 379 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2651fc120>  
 parent: proto object 
 .. parent: proto object 
 .. .. parent: proto object 
}}
\usage{
BooleanEditor
}
\description{
Editor has regular or compact style
}
\keyword{datasets}

