% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Container}
\alias{Container}
\title{Base Trait for Container objects. Containers are used to make views.}
\format{\preformatted{proto object 
 $ attr          :List of 1 
 $ class         : chr [1:2] "Container" "TraitR" 
 $ make_ui       :function (., container, attr = .$attr, context, ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 124 40 171 31 40 31 124 171 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c4e4a0>  
 $ init_ui       :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 115 38 115 51 38 51 115 115 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c4e4a0>  
 $ context       : NULL 
 $ validate_ui   :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 199 43 199 56 43 56 199 199 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c4e4a0>  
 $ is_realized   :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 48 43 48 100 43 100 48 48 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c4e4a0>  
 $ type          : chr "glayout" 
 $ visible_when  :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 190 46 190 65 46 65 190 190 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c4e4a0>  
 $ instance      :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 236 39 248 30 39 30 236 248 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c4e4a0>  
 $ enabled       :function (., value)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 182 41 184 31 41 31 182 184 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c4e4a0>  
 $ no_cols       : num 1 
 $ update_ui     :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 204 43 232 31 43 31 204 232 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c4e4a0>  
 $ make_container:function (., container, attr = .$attr)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 50 46 111 31 46 31 50 111 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c4e4a0>  
 $ enabled_when  :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 177 46 180 31 46 31 177 180 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c4e4a0>  
 $ children      : list() 
 $ container     : NULL 
 $ visible       :function (., value)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 192 41 194 31 41 31 192 194 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c4e4a0>  
 parent: proto object 
}}
\usage{
Container
}
\description{
Basic container is a glayout object for tabular layout
There are various types of layouts. The most basic, and default view, is simply
\code{aContainer(...names of items...)} which simply uses a table to display the item's label
and editor. Other containers can be used to adjust this.
}
\details{
Containers have a few methods, notably the \code{is_visible} and
\code{is_enabled} methods which can be used to hide or set
sensitive to  user input the container's components.
}
\keyword{datasets}

