% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Controller}
\alias{Controller}
\title{Trait for Controller objects}
\format{\preformatted{proto object 
 $ get_model        :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 43 42 43 60 42 60 43 43 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266cf1948>  
 $ remove_view      :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 76 44 78 32 44 32 76 78 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266cf1948>  
 $ class            : chr [1:2] "Controller" "TraitR" 
 $ view             : NULL 
 $ update_from_model:function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 85 50 85 63 50 63 85 85 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266cf1948>  
 $ get_view         :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 61 41 61 58 41 58 61 61 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266cf1948>  
 $ set_model        :function (., model)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 48 44 55 32 44 32 48 55 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266cf1948>  
 $ adapters         : list() 
 $ init             :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 119 39 144 32 39 32 119 144 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266cf1948>  
 $ model            : NULL 
 $ register_adapters:function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 97 52 112 32 52 32 97 112 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266cf1948>  
 $ update_from_view :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 91 51 91 64 51 64 91 91 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266cf1948>  
 $ set_view         :function (., view)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 67 43 74 32 43 32 67 74 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266cf1948>  
 parent: proto object 
}}
\usage{
Controller
}
\description{
A controller connects a model and an associated view to synchronize changes in one with another
This implementation rests on the controller having some suitably named methods
}
\keyword{datasets}

