% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dfEditItem}
\alias{dfEditItem}
\title{data frame editor item. Needs writing}
\usage{
dfEditItem(...)
}
\arguments{
\item{...}{to be replaced with actual arguments}
}
\value{
A \code{proto} object. Call \code{obj$show_help()} to view its methods and properties.
}
\description{
Write me
}

