% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Dialog}
\alias{Dialog}
\title{A Dialog wraps a top-level window around a collection of items which may be item groups}
\format{\preformatted{proto object 
 $ make_gui          :function (., gui_layout = .$make_default_gui_layout(), parent = NULL, visible = TRUE)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 127 36 214 27 36 27 127 214 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2668d25f8>  
 $ OK_handler        :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 80 38 82 27 38 27 80 82 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2668d25f8>  
 $ Redo_handler      :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 102 40 102 59 40 59 102 102 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2668d25f8>  
 $ doc_Cancel_handler: chr "<p class=description>Description: Handler called when 'Cancel' button is clicked -- if 'Cancel' button among <code>buttons</cod"| __truncated__ 
 $ default_button    : NULL 
 $ update_ui         :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 233 37 243 27 37 27 233 243 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2668d25f8>  
 $ doc_OK_handler    : chr "<p class=description>Description: Handler called when 'OK' button is clicked -- if 'OK' button among <code>buttons</code>proper"| __truncated__ 
 $ items             : list() 
 $ status_text       : NULL 
 $ doc_default_button: chr "<p class=description>Description: Name of default button. Leave empty (<code>NULL</code>) for none.</p>" 
 $ Undo_handler      :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 98 40 98 59 40 59 98 98 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2668d25f8>  
 $ Help_handler      :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 92 42 94 27 42 27 92 94 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2668d25f8>  
 $ class             : chr [1:4] "Dialog" "ItemGroup" "Model" "TraitR" 
 $ toolbar_list      : NULL 
 $ on_realized       :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 104 41 106 27 41 27 104 106 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2668d25f8>  
 $ visible           :function (., value = TRUE)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 228 35 231 27 35 27 228 231 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2668d25f8>  
 $ title             : chr "Dialog title" 
 $ close_gui         :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 219 39 222 27 39 27 219 222 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2668d25f8>  
 $ doc_Help_handler  : chr "<p class=description>Description: Handler called when 'Help' button is clicked -- if 'Help' button among <code>buttons</code>pr"| __truncated__ 
 $ doc_Redo_handler  : chr "<p class=description>Description: Handler called when 'Redo' button is clicked -- if 'Redo' button among <code>buttons</code>pr"| __truncated__ 
 $ Cancel_handler    :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 86 44 88 27 44 27 86 88 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2668d25f8>  
 $ doc_Undo_handler  : chr "<p class=description>Description: Handler called when 'Undo' button is clicked -- if 'Undo' button among <code>buttons</code>pr"| __truncated__ 
 $ buttons           : chr [1:4] "OK" "Cancel" "SPACE" "Help" 
 $ get_widget        :function (., key)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 112 38 112 73 38 73 112 112 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2668d25f8>  
 $ widget_list       : list() 
 $ menu_list         : NULL 
 $ help_string       : chr "" 
 parent: proto object 
 .. parent: proto object 
 .. .. parent: proto object 
}}
\usage{
Dialog
}
\description{
One can specify the parent when making the UI
Buttons are specified through the buttons property
}
\keyword{datasets}

