% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Editor}
\alias{ButtonEditor}
\alias{DateEditor}
\alias{Editor}
\alias{EntryEditor}
\alias{FileBrowseEditor}
\alias{GraphEditor}
\alias{ImageEditor}
\alias{LabelEditor}
\alias{ObjectWithValuesEditor}
\alias{RangeEditor}
\alias{SeparatorEditor}
\alias{TableEditor}
\title{Base Trait for Editor.}
\format{\preformatted{proto object 
 $ attr               :List of 1 
 $ class              : chr [1:3] "Editor" "View" "TraitR" 
 $ make_ui            :function (., container, attr = .$attr, context = ., ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 52 32 66 22 32 22 52 66 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2651fc120>  
 $ editor_name        : chr "gedit" 
 $ set_valid          :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 187 32 187 45 32 45 187 187 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2651fc120>  
 $ instance           :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 179 31 183 22 31 22 179 183 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2651fc120>  
 $ enabled            :function (., value)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 134 32 136 22 32 22 134 136 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2651fc120>  
 $ view_widget_name   : chr "editor" 
 $ editor_style       : NULL 
 $ set_value_in_view  :function (., widget_name = .$view_widget_name, value)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 162 42 174 22 42 22 162 174 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2651fc120>  
 $ set_invalid        :function (., mesg)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 191 34 194 22 34 22 191 194 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2651fc120>  
 $ get_value_from_view:function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 142 44 152 22 44 22 142 152 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2651fc120>  
 $ visible            :function (., value)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 131 32 133 22 32 22 131 133 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc2651fc120>  
 parent: proto object 
 .. parent: proto object 
}}
\usage{
Editor

EntryEditor

ObjectWithValuesEditor

RangeEditor

ButtonEditor

ImageEditor

GraphEditor

FileBrowseEditor

DateEditor

TableEditor

LabelEditor

SeparatorEditor
}
\description{
An editor is a basic view for a widget, essentially a map from gedit, say, to a view.

A Base Trait for an editor using the entry widget

editor has regular or compact style

Trait for making a range editor (slider, spinbox)

Trait for button editor

Trait for embedding an image file

Trait for embedding graphics (Qt, RGtk2 only)

Trait for making File browser editor

Trait for data selection editor

No selection, just display. For selection use ChoiceItem

Trait for a label

Trait for making a visual separator
}
\keyword{datasets}

