% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_with_default}
\alias{get_with_default}
\title{Function to return value or an object (or default if value is null, NA or "")}
\usage{
get_with_default(x, default)
}
\arguments{
\item{x}{object}

\item{default}{default value}
}
\value{
Returns default if x is NA, null or "", otherwise x
}
\description{
Function to return value or an object (or default if value is null, NA or "")
}

