% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDB.R
\name{getPytaxonSpeciesName}
\alias{getPytaxonSpeciesName}
\title{Returns a Data frame with the species names curated by Pytaxon. Pytaxon results are saved in the temp folder of your project.}
\usage{
getPytaxonSpeciesName(speciesNames)
}
\arguments{
\item{speciesNames}{String. Comma separated list of scientific name for species to analyze through Pytaxon.}
}
\value{
A data frame with the names of the species returned by Pytaxon. Entries are in the same order as the input list.
}
\description{
Returns a Data frame with the species names curated by Pytaxon. Pytaxon results are saved in the temp folder of your project.
}
