% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ItemListEditor}
\alias{ItemListEditor}
\title{trait for editor for itemList}
\format{\preformatted{proto object 
 $ update_ui        :function (., context)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 350 44 370 32 44 32 350 370 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc265e2c778>  
 $ class            : chr [1:4] "ItemListEditor" "Editor" "View" "TraitR" 
 $ get_table_entries:function (., items, name)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 219 50 232 32 50 32 219 232 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc265e2c778>  
 $ clear_edit_area  :function (., message = "")   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 235 50 241 32 50 32 235 241 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc265e2c778>  
 $ make_ui          :function (., container, attr = .attr, context = ., ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 269 42 348 32 42 32 269 348 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc265e2c778>  
 $ edit_item        :function (., context, item)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 243 44 267 32 44 32 243 267 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc265e2c778>  
 parent: proto object 
 .. parent: proto object 
 .. .. parent: proto object 
}}
\usage{
ItemListEditor
}
\description{
trait for editor for itemList
}
\keyword{datasets}

