\name{merge.list}
\alias{merge.list}
\title{merge two lists, possibly overwriting...}
\usage{merge.list(x, y, overwrite=TRUE)}
\description{merge two lists, possibly overwriting}
\value{Returns a list}
\arguments{\item{x}{a list}
\item{y}{a list. Named values of \code{y} are assigned to \code{x} and then \code{x} is returned.}
\item{overwrite}{If \code{TRUE} named values in \code{y} clobber similarly named values in \code{x}}}
