% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wrap_in_tag}
\alias{desc}
\alias{loadingAnimation}
\alias{merge.list}
\alias{param}
\alias{returns}
\alias{ul}
\alias{wrap_in_tag}
\title{write values in paired tags with optional class}
\usage{
wrap_in_tag(tag, ..., class = "")

desc(...)

param(value, ...)

returns(...)

ul(values)

\method{merge}{list}(x, y, overwrite = TRUE)

loadingAnimation(message = "<b>Loading...</b>")
}
\arguments{
\item{tag}{tag to wrap in, eg. "ul"}

\item{...}{values to be wrapped. Pasted together}

\item{class}{optional call to add to tag}

\item{value}{name of parameter
return a string}

\item{values}{values to put into list items}

\item{x}{a list}

\item{y}{a list. Named values of \code{y} are assigned to \code{x} and then \code{x} is returned.}

\item{overwrite}{If \code{TRUE} named values in \code{y} clobber similarly named values in \code{x}}

\item{message}{A message to display along with graphic while loading. PANGO markup is okay.}
}
\value{
a string

returns a string

Returns a list

An item group instance with a \code{close} method to call to dismiss window
}
\description{
internal function for writing proto docs

internal function for writing proto docs.
return a string

internal function for writing proto docs.

internal function for writing proto docs.
return a string

internal function for writing proto docs.

merge two lists, possibly overwriting

A window to show a loading animation
}
\examples{
## we call, something happens, then we close
\dontrun{
w <- loadingAnimation()
## .... something long, like dlg$make_gui() ...
w$close()
}
}

