\name{separatorItem}
\alias{separatorItem}
\title{Visual separator item...}
\usage{separatorItem(name=".seperator", attr=list(horizontal = TRUE), ...)}
\description{Visual separator item}
\value{A \code{proto} object. Call \code{obj$show_help()} to view its methods and properties.}
\arguments{\item{name}{name of widget}
\item{attr}{passed to widget constructor. Use list(horizontal=FALSE) to get vertical}
\item{...}{ignored}}
