% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDB.R
\name{speciesInTrees}
\alias{speciesInTrees}
\title{Returns a Data frame with the species names included in the provided list of trees.}
\usage{
speciesInTrees(treeNames)
}
\arguments{
\item{treeNames}{String. Comma separated list of trees to use.}
}
\value{
A data frame with the species included in the provided trees.
}
\description{
Returns a Data frame with the species names included in the provided list of trees.
}
