% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDB.R
\name{speciesTreeName}
\alias{speciesTreeName}
\title{Returns a Data frame where each entry is ";" separated tupple containing: first the name of the species as it appears in the tree the species names as are in the tree and second the name
of the species used in the database.}
\usage{
speciesTreeName(treeName)
}
\arguments{
\item{treeName}{String. Tree to obtain species names from.}
}
\value{
A data frame where each entry is ";" separated tupple containing: first the name of the species as it appears in the tree the species names as are in the tree and second the name
of the species used in the database.
}
\description{
Returns a Data frame where each entry is ";" separated tupple containing: first the name of the species as it appears in the tree the species names as are in the tree and second the name
of the species used in the database.
}
