% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDB.R
\name{traitSearch}
\alias{traitSearch}
\title{Obtain requested trait data for specified species from specified datasets.}
\usage{
traitSearch(speciesName, traits, datasets)
}
\arguments{
\item{speciesName}{String. Scientific name of the species you are interested in obtaining data for.}

\item{traits}{String. Comma separated list of traits you want the data for.}

\item{datasets}{String. Comma separated list of datasets containing the desired traits.}
}
\value{
Data frame with the desired trait data for the specified species. Function will also save results to out/testObtain.csv.
}
\description{
Obtain requested trait data for specified species from specified datasets.
}
\examples{

traitSearch(speciesHacketString,traitsNestString,"Nest data")

}
