% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDB.R
\name{traitsInDatasets}
\alias{traitsInDatasets}
\title{Returns a Data frame with the names of the columns (traits) included in the provided list of datasets.}
\usage{
traitsInDatasets(datasetNames)
}
\arguments{
\item{datasetName}{String. Comma separated list of dataset names.}
}
\value{
A data frame with the traits included in the provided datasets.
}
\description{
Returns a Data frame with the names of the columns (traits) included in the provided list of datasets.
}
