% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDB.R
\name{treesContainSpecies}
\alias{treesContainSpecies}
\title{Obtain the trees that contain the provided species.}
\usage{
treesContainSpecies(speciesNames)
}
\arguments{
\item{speciesNames}{String. Comma separated list of scientific name for species to find.}
}
\value{
A data frame with the trees that contain at least one of the specified species.
}
\description{
Function returns a Data frame that lists all trees that contain the species provided as input.
}
\details{
The function looks for each species in the master list and builds a list of unique tree name that contain at least one of the species requested.

The result is a list of all trees that contain at least one of the species requested.
}
\examples{

speciesHacket <- speciesInTrees("origin/Stage2_Hackett_MCC_no_neg.tre")
speciesHacketString <- paste(speciesHacket,collapse=",")
treesContainSpecies(speciesHacketString)

}
