% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{View}
\alias{View}
\title{Trait for View objects.}
\format{\preformatted{proto object 
 $ class              : chr [1:2] "View" "TraitR" 
 $ is_realized        :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 62 38 64 26 38 26 62 64 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c784a0>  
 $ set_value_in_view  :function (., widget_name, value)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 79 45 87 25 45 25 79 87 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c784a0>  
 $ widgets            : list() 
 $ attr               : list() 
 $ get_value_from_view:function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 73 47 73 60 47 60 73 73 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c784a0>  
 $ enabled            :function (., value)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 94 36 96 26 36 26 94 96 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c784a0>  
 $ get_widgets        :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 43 39 43 59 39 59 43 43 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c784a0>  
 $ visible            :function (., value)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 101 36 103 26 36 26 101 103 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c784a0>  
 $ make_ui            :function (., cont, attr = .$attr)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 57 34 57 66 34 66 57 57 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c784a0>  
 $ get_widget_by_name :function (., key)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 48 45 48 83 45 83 48 48 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fc266c784a0>  
 parent: proto object 
}}
\usage{
View
}
\description{
A view "displays" the values in an associated model. The association is through a controller
the view does not know the controller or the model
Views are initialized through the \code{make_ui} method.
}
\keyword{datasets}

