\name{wrap_in_tag}
\alias{wrap_in_tag}
\title{write values in paired tags with optional class...}
\usage{wrap_in_tag(tag, ..., class="")}
\description{write values in paired tags with optional class}
\details{internal function for writing proto docs}
\value{a string}
\arguments{\item{tag}{tag to wrap in, eg. "ul"}
\item{...}{values to be wrapped. Pasted together}
\item{class}{optional call to add to tag}}
