\name{aController}
\alias{aController}
\title{Constructor for a Controller proto objects...}
\usage{aController(...)}
\description{Constructor for a Controller proto objects}
\details{Simply provides a more typical calling interface for the Controller proto object}
\value{returns the Controller object}
\arguments{\item{...}{passed to proto method for}}
