% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDB.R
\name{addDataset}
\alias{addDataset}
\title{Function to add a new dataset to the local database.}
\usage{
addDataset(datasetFile, datasetFileType, datasetName, speciesColumn)
}
\arguments{
\item{datasetFile}{String. Relative path of the file containing the dataset you wish to add.}

\item{datasetFileType}{String. The format of the dataset file. Accepts: excel, csv.}

\item{datasetName}{String. The name of the dataset.}

\item{speciesColumn}{String. The name of the column that contains the scientific names of the species in the dataset.}
}
\description{
Function to add a new dataset to the local database.
}
