% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDB.R
\name{addTree}
\alias{addTree}
\title{Add a new phylogenetic tree to the comparative database.}
\usage{
addTree(treeFile, treeFileFormat)
}
\arguments{
\item{treeFile}{String. Relative path, from execution path, of the file containing the tree you wish to add.}

\item{treeFileFormat}{String. The format of the dataset file. Accepts: newick, nexus, nexml, phyloxml, cdao.}
}
\description{
Adds the provided phylogenetic tree to the comparative database and regiters the species it contains.
}
\details{
The function loads the provided tree file (see below for accepted format) register it in the trees.csv file and registers which species are in the tree.

The function also runs the species names in the tree through Pytaxon and then looks them up in the system's master list. If the species is found it registers
that the species can be found in this tree, if it is not found it adds the species to the verification list so that a user can later match it to
the appropriate species in the master list.

The tree formats accepted are: newick, nexus, nexml, phyloxml and cdao
}
\examples{
In this example we will add the tree "Stage2_Hackett_MCC_no_neg.tre" to the database. This tree is in newick format, the file is located in the folder
"origin". To add it we would run the following instruction:

addTree("origin/Stage2_Hackett_MCC_no_neg.tre","newick")
}
