\name{aModel}
\alias{aModel}
\title{Constructor for a Model proto objects...}
\usage{aModel(...)}
\description{Constructor for a Model proto objects}
\details{Simply provides a more typical calling interface for the Model proto object}
\value{returns the Model object. Call \code{obj$show_help()} to view its methods and properties.}
\arguments{\item{...}{passed to proto method for}}
