\name{aView}
\alias{aView}
\title{Constructor for a View proto object...}
\usage{aView(...)}
\description{Constructor for a View proto object}
\details{Simply provides a more typical calling interface for the View proto object}
\value{Returns the View object. Call \code{obj$show_help()} to view its methods and properties.}
\arguments{\item{...}{passed to proto method for}}
