\name{Controller}
\alias{Controller}
\title{Trait for Controller objects...}
\description{Trait for Controller objects}
\details{A controller connects a model and an associated view to synchronize changes in one with another
This implementation rests on the controller having some suitably named methods}
