% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDB.R
\name{createDatabase}
\alias{createDatabase}
\title{Setups the initial database that will be the foundation of the system.
Please ensure that your project has a data, temp and out folders.
Please ensure you have the config.json and AviList-v2025-11Jun-extended.xlsx files found in the github in the
root of your project.
Run this only for initial setup, if you run it again it will overwrite the main database file, which will cause the system to break.}
\usage{
createDatabase()
}
\description{
Setups the initial database that will be the foundation of the system.
Please ensure that your project has a data, temp and out folders.
Please ensure you have the config.json and AviList-v2025-11Jun-extended.xlsx files found in the github in the
root of your project.
Run this only for initial setup, if you run it again it will overwrite the main database file, which will cause the system to break.
}
