% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDB.R
\name{datasetsIncludeSpecies}
\alias{datasetsIncludeSpecies}
\title{Returns a Data frame with all datasets that contain the provided species.}
\usage{
datasetsIncludeSpecies(speciesName)
}
\arguments{
\item{speciesName}{String. Comma separated list of scientific name for species to find.}
}
\value{
Data frame with all datasets in database that contain at least one of the species provided.
}
\description{
Returns a Data frame with all datasets that contain the provided species.
}
