\name{dfEditItem}
\alias{dfEditItem}
\title{data frame editor item.}
\usage{dfEditItem(...)}
\description{data frame editor item. Needs writing}
\details{Write me}
\value{A \code{proto} object. Call \code{obj$show_help()} to view its methods and properties.}
\arguments{\item{...}{to be replaced with actual arguments}}
