\name{Dialog}
\alias{Dialog}
\title{A Dialog wraps a top-level window around a collection of items which may be item groups...}
\description{A Dialog wraps a top-level window around a collection of items which may be item groups}
\details{One can specify the parent when making the UI
Buttons are specified through the buttons property}
