\name{Model}
\alias{Model}
\title{Trait for a model object.}
\description{Trait for a model object.}
\details{Model objects consist of properties and the methods that manipulate them
Models are initialized by \code{init} so that getter and setter pairs are made
When setter functions are called, the model's observers are notified}
