\name{View}
\alias{View}
\title{Trait for View objects.}
\description{Trait for View objects.}
\details{A view "displays" the values in an associated model. The association is through a controller
the view does not know the controller or the model
Views are initialized through the \code{make_ui} method.}
